moves some functions to utils.h and stops exporting utils in cpp
